- add routing strategies for http  - add streaming example from router to client
